+++ MonsterDB/Managers/Item/Components/ItemDataSharedRef.cs
@@ -18,44 +18,44 @@ public sealed class ItemDataSharedRef : Reference
     [Condition(icon = true)] public string[]? m_icons;
     [Condition(icon = true)] public ItemDrop.ItemData.ItemType? m_attachOverride;
     [Condition(icon = true)] public string? m_description;
-    [Condition(icon = true)][DefaultValue(1)] public int? m_maxStackSize;
-    [Condition(icon = true)][DefaultValue(true)] public bool? m_autoStack;
-    [Condition(icon = true)][DefaultValue(1)] public int? m_maxQuality;
-    [Condition(icon = true)][DefaultValue(0f)] public float? m_scaleByQuality;
-    [Condition(icon = true)][DefaultValue(0f)] public float? m_weight;
-    [Condition(icon = true)][DefaultValue(0f)] public float? m_scaleWeightByQuality;
-    [Condition(icon = true)][DefaultValue(0)] public int? m_value;
+    [Condition(icon = true)] public int? m_maxStackSize;
+    [Condition(icon = true)] public bool? m_autoStack;
+    [Condition(icon = true)] public int? m_maxQuality;
+    [Condition(icon = true)] public float? m_scaleByQuality;
+    [Condition(icon = true)] public float? m_weight;
+    [Condition(icon = true)] public float? m_scaleWeightByQuality;
+    [Condition(icon = true)] public int? m_value;
     [Condition(icon = true)] public bool? m_teleportable;
-    [Condition(icon = true)][DefaultValue(false)] public bool? m_questItem;
-    [Condition(icon = true)][DefaultValue(0f)] public float? m_equipDuration;
-    [Condition(icon = true)][DefaultValue(0)] public int? m_variants;
+    [Condition(icon = true)] public bool? m_questItem;
+    [Condition(icon = true)] public float? m_equipDuration;
+    [Condition(icon = true)] public int? m_variants;
     
     [YamlMember(Description = "Set Settings")]
-    [Condition(icon = true)][DefaultValue(0)] public int? m_setSize;
+    [Condition(icon = true)] public int? m_setSize;
     [Condition(icon = true)] public string? m_setStatusEffect;
     
     [Condition(icon = true)] public string? m_equipStatusEffect;
 
     [YamlMember(Description = "Modifiers")] 
-    [Condition(icon = true)][DefaultValue(0f)] public float? m_eitrRegenModifier;
-    [Condition(icon = true)][DefaultValue(0f)] public float? m_movementModifier;
-    [Condition(icon = true)][DefaultValue(0f)] public float? m_homeItemsStaminaModifier;
-    [Condition(icon = true)][DefaultValue(0f)] public float? m_heatResistanceModifier;
-    [Condition(icon = true)][DefaultValue(0f)] public float? m_jumpStaminaModifier;
-    [Condition(icon = true)][DefaultValue(0f)] public float? m_attackStaminaModifier;
-    [Condition(icon = true)][DefaultValue(0f)] public float? m_blockStaminaModifier;
-    [Condition(icon = true)][DefaultValue(0f)] public float? m_dodgeStaminaModifier;
-    [Condition(icon = true)][DefaultValue(0f)] public float? m_swimStaminaModifier;
-    [Condition(icon = true)][DefaultValue(0f)] public float? m_sneakStaminaModifier;
-    [Condition(icon = true)][DefaultValue(0f)] public float? m_runStaminaModifier;
+    [Condition(icon = true)] public float? m_eitrRegenModifier;
+    [Condition(icon = true)] public float? m_movementModifier;
+    [Condition(icon = true)] public float? m_homeItemsStaminaModifier;
+    [Condition(icon = true)] public float? m_heatResistanceModifier;
+    [Condition(icon = true)] public float? m_jumpStaminaModifier;
+    [Condition(icon = true)] public float? m_attackStaminaModifier;
+    [Condition(icon = true)] public float? m_blockStaminaModifier;
+    [Condition(icon = true)] public float? m_dodgeStaminaModifier;
+    [Condition(icon = true)] public float? m_swimStaminaModifier;
+    [Condition(icon = true)] public float? m_sneakStaminaModifier;
+    [Condition(icon = true)] public float? m_runStaminaModifier;
 
     [YamlMember(Description = "Food Settings")]
-    [Condition(icon = true)][DefaultValue(0f)] public float? m_food;
-    [Condition(icon = true)][DefaultValue(0f)] public float? m_foodStamina;
-    [Condition(icon = true)][DefaultValue(0f)] public float? m_foodEitr;
-    [Condition(icon = true)][DefaultValue(0f)] public float m_foodBurnTime;
-    [Condition(icon = true)][DefaultValue(0f)] public float? m_foodRegen;
-    [Condition(icon = true)][DefaultValue(false)] public bool? m_isDrink;
+    [Condition(icon = true)] public float? m_food;
+    [Condition(icon = true)] public float? m_foodStamina;
+    [Condition(icon = true)] public float? m_foodEitr;
+    [Condition(icon = true)] public float m_foodBurnTime;
+    [Condition(icon = true)] public float? m_foodRegen;
+    [Condition(icon = true)] public bool? m_isDrink;
 
     [YamlMember(Description = "Armor Settings")]
     [Condition(icon = true)] public string? m_armorMaterial;
@@ -85,15 +85,15 @@ public sealed class ItemDataSharedRef : Reference
     [YamlMember(Description = "Weapon")] 
     public ItemDrop.ItemData.AnimationState? m_animationState;
     [Condition(icon = true)] public Skills.SkillType? m_skillType;
-    [DefaultValue(0)] public int? m_toolTier;
+    public int? m_toolTier;
     public HitData.DamageTypes? m_damages;
-    [DefaultValue(0f)] public float? m_attackForce;
-    [DefaultValue(0f)] public float? m_backstabBonus;
-    [DefaultValue(false)] public bool? m_dodgeable;
-    [DefaultValue(false)] public bool? m_blockable;
+    public float? m_attackForce;
+    public float? m_backstabBonus;
+    public bool? m_dodgeable;
+    public bool? m_blockable;
     [Condition] public bool? m_tamedOnly;
     public string? m_attackStatusEffect;
-    [DefaultValue(1f)] public float? m_attackStatusEffectChance;
+    public float? m_attackStatusEffectChance;
     public string? m_spawnOnHit;
     public string? m_spawnOnHitTerrain;
     public string? m_ammoType;
@@ -106,19 +106,19 @@ public sealed class ItemDataSharedRef : Reference
     [Condition(icon = true)] public float? m_durabilityPerLevel;
     [Condition(icon = true)] public float m_useDurabilityDrain;
     [YamlMember(Description = "AI behaviour")]
-    [DefaultValue(0f)] public float? m_aiAttackRange;
-    [DefaultValue(0f)] public float? m_aiAttackRangeMin;
-    [DefaultValue(0f)] public float? m_aiAttackInterval;
-    [DefaultValue(0f)] public float? m_aiAttackMaxAngle;
-    [DefaultValue(false)] public bool? m_aiInvertAngleCheck;
-    [DefaultValue(false)] public bool? m_aiWhenFlying;
-    [DefaultValue(0f)] public float? m_aiWhenFlyingAltitudeMin;
-    [DefaultValue(999999f)] public float? m_aiWhenFlyingAltitudeMax;
-    [DefaultValue(false)] public bool? m_aiWhenWalking;
-    [DefaultValue(false)] public bool? m_aiWhenSwiming;
-    [DefaultValue(false)] public bool? m_aiPrioritized;
-    [DefaultValue(false)] public bool? m_aiInDungeonOnly;
-    [DefaultValue(false)] public bool? m_aiInMistOnly;
+    public float? m_aiAttackRange;
+    public float? m_aiAttackRangeMin;
+    public float? m_aiAttackInterval;
+    public float? m_aiAttackMaxAngle;
+    public bool? m_aiInvertAngleCheck;
+    public bool? m_aiWhenFlying;
+    public float? m_aiWhenFlyingAltitudeMin;
+     public float? m_aiWhenFlyingAltitudeMax;
+    public bool? m_aiWhenWalking;
+    public bool? m_aiWhenSwiming;
+    public bool? m_aiPrioritized;
+    public bool? m_aiInDungeonOnly;
+    public bool? m_aiInMistOnly;
     [DefaultValue(1f)] public float? m_aiMaxHealthPercentage;
     [DefaultValue(0f)]public float? m_aiMinHealthPercentage;
     public ItemDrop.ItemData.AiTarget? m_aiTargetType;