+++ MonsterDB/Managers/Ragdoll/RagdollRef.cs
@@ -6,7 +6,7 @@ namespace MonsterDB;
 [Serializable]
 public class RagdollRef : Reference
 {
-    [DefaultValue(1f)] public float? m_velMultiplier;
+    public float? m_velMultiplier;
     public float? m_ttl;
     public EffectListRef? m_removeEffect;
     public bool? m_float;